Partially revert commit 0e41b1911
This commit added "Workaround for TSAN false positive reports with
glib". With properly compiled glib-2.0 libraries for TSAN check such
workaround is no longer required.
