+++ src/platform/Linux/PlatformManagerImpl.cpp
@@ -33,7 +33,9 @@
 #include <netinet/in.h>
 #include <unistd.h>
 
-#include <mutex>
+#if CHIP_DEVICE_CONFIG_WITH_GLIB_MAIN_LOOP
+#include <glib.h>
+#endif
 
 #include <app-common/zap-generated/enums.h>
 #include <app-common/zap-generated/ids/Events.h>
@@ -56,12 +58,47 @@ PlatformManagerImpl PlatformManagerImpl::sInstance;
 namespace {
 
 #if CHIP_DEVICE_CONFIG_WITH_GLIB_MAIN_LOOP
+
+class CallbackIndirection
+{
+public:
+    CallbackIndirection(GSourceFunc callback, void * userData) : mCallback(callback), mUserData(userData) {}
+
+    void Wait()
+    {
+        g_mutex_lock(&mDoneMtx);
+        while (!mDone)
+            g_cond_wait(&mDoneCond, &mDoneMtx);
+        g_mutex_unlock(&mDoneMtx);
+    }
+
+    static gboolean Callback(CallbackIndirection * self)
+    {
+        auto result = self->mCallback(self->mUserData);
+
+        g_mutex_lock(&self->mDoneMtx);
+        self->mDone = true;
+        g_mutex_unlock(&self->mDoneMtx);
+
+        g_cond_signal(&self->mDoneCond);
+        return result;
+    }
+
+private:
+    GMutex mDoneMtx{};
+    GCond mDoneCond{};
+    bool mDone = false;
+    GSourceFunc mCallback;
+    void * mUserData;
+};
+
 void * GLibMainLoopThread(void * loop)
 {
     g_main_loop_run(static_cast<GMainLoop *>(loop));
     return nullptr;
 }
-#endif
+
+#endif // CHIP_DEVICE_CONFIG_WITH_GLIB_MAIN_LOOP
 
 #if CHIP_DEVICE_CONFIG_ENABLE_WIFI
 
@@ -193,12 +230,9 @@ CHIP_ERROR PlatformManagerImpl::_InitChipStack()
     mGLibMainLoop       = g_main_loop_new(nullptr, FALSE);
     mGLibMainLoopThread = g_thread_new("gmain-matter", GLibMainLoopThread, mGLibMainLoop);
 
-    {
-        std::unique_lock<std::mutex> lock(mGLibMainLoopCallbackIndirectionMutex);
-        CallbackIndirection startedInd([](void *) { return G_SOURCE_REMOVE; }, nullptr);
-        g_idle_add(G_SOURCE_FUNC(&CallbackIndirection::Callback), &startedInd);
-        startedInd.Wait(lock);
-    }
+    CallbackIndirection startedInd([](void *) { return G_SOURCE_REMOVE; }, nullptr);
+    g_idle_add(G_SOURCE_FUNC(&CallbackIndirection::Callback), &startedInd);
+    startedInd.Wait();
 
 #endif
 
@@ -253,34 +287,7 @@ void PlatformManagerImpl::_Shutdown()
 #endif
 }
 
-#if CHIP_DEVICE_CONFIG_WITH_GLIB_MAIN_LOOP
-
-void PlatformManagerImpl::CallbackIndirection::Wait(std::unique_lock<std::mutex> & lock)
-{
-    mDoneCond.wait(lock, [this]() { return mDone; });
-}
-
-gboolean PlatformManagerImpl::CallbackIndirection::Callback(CallbackIndirection * self)
-{
-    // We can not access "self" before acquiring the lock, because TSAN will complain that
-    // there is a race condition between the thread that created the object and the thread
-    // that is executing the callback.
-    std::unique_lock<std::mutex> lock(PlatformMgrImpl().mGLibMainLoopCallbackIndirectionMutex);
-
-    auto callback = self->mCallback;
-    auto userData = self->mUserData;
-
-    lock.unlock();
-    auto result = callback(userData);
-    lock.lock();
-
-    self->mDone = true;
-    self->mDoneCond.notify_all();
-
-    return result;
-}
-
-#if CHIP_DEVICE_CONFIG_ENABLE_CHIPOBLE
+#if CHIP_DEVICE_CONFIG_WITH_GLIB_MAIN_LOOP && CHIP_DEVICE_CONFIG_ENABLE_CHIPOBLE
 CHIP_ERROR PlatformManagerImpl::RunOnGLibMainLoopThread(GSourceFunc callback, void * userData, bool wait)
 {
 
@@ -288,29 +295,18 @@ CHIP_ERROR PlatformManagerImpl::RunOnGLibMainLoopThread(GSourceFunc callback, vo
     VerifyOrReturnError(context != nullptr, CHIP_ERROR_INTERNAL,
                         ChipLogDetail(DeviceLayer, "Failed to get GLib main loop context"));
 
-    // If we've been called from the GLib main loop thread itself, there is no reason to wait
-    // for the callback, as it will be executed immediately by the g_main_context_invoke() call
-    // below. Using a callback indirection in this case would cause a deadlock.
-    if (g_main_context_is_owner(context))
-    {
-        wait = false;
-    }
-
     if (wait)
     {
-        std::unique_lock<std::mutex> lock(mGLibMainLoopCallbackIndirectionMutex);
         CallbackIndirection indirection(callback, userData);
         g_main_context_invoke(context, G_SOURCE_FUNC(&CallbackIndirection::Callback), &indirection);
-        indirection.Wait(lock);
+        indirection.Wait();
         return CHIP_NO_ERROR;
     }
 
     g_main_context_invoke(context, callback, userData);
     return CHIP_NO_ERROR;
 }
-#endif // CHIP_DEVICE_CONFIG_ENABLE_CHIPOBLE
-
-#endif // CHIP_DEVICE_CONFIG_WITH_GLIB_MAIN_LOOP
+#endif
 
 } // namespace DeviceLayer
 } // namespace chip
